Yonatan/136/Move all practices features to a new folder named practice (#141)
