+++ frontend/src/features/index.ts
@@ -1,5 +1,3 @@
 export * from "./avatar";
 export * from "./chat";
-export * from "./typing-practice";
-export * from "./word-order-game";
-export * from "./speaking-practice";
+export * from "./practice";
