+++ frontend/src/features/practice/index.ts
@@ -1,0 +1,3 @@
+export * from "./speaking-practice";
+export * from "./typing-practice";
+export * from "./word-order-game";
