+++ frontend/src/features/practice/typing-practice/components/AudioControls/index.tsx
