+++ frontend/src/features/practice/typing-practice/components/AudioControls/style.ts
