+++ frontend/src/features/practice/typing-practice/components/FeedbackDisplay/style.ts
