+++ frontend/src/features/practice/typing-practice/components/index.ts
