+++ frontend/src/features/practice/typing-practice/style.ts
