+++ frontend/src/features/practice/typing-practice/types.ts
