+++ frontend/src/features/practice/typing-practice/utils.ts
