+++ frontend/src/features/practice/word-order-game/components/Description/index.tsx
