+++ frontend/src/features/practice/word-order-game/components/Game/style.ts
