+++ frontend/src/features/practice/word-order-game/components/Speaker/index.tsx
