+++ frontend/src/features/practice/word-order-game/word-order-game.stories.tsx
@@ -2,12 +2,12 @@ import { WordOrderGame } from "./index";
 import type { Meta, StoryObj } from "@storybook/react-vite";
 
 const meta: Meta = {
-    title: "Features/WordOrderGame",
-    component: WordOrderGame,
+  title: "Features/WordOrderGame",
+  component: WordOrderGame,
 };
 
 export default meta;
 
 export const Default: StoryObj = {
-    render: () => <WordOrderGame />,
+  render: () => <WordOrderGame />,
 };
